update: ワークスペースの導入: coreモジュールに定義していた`Parser`をwasmモジュールに移動
