+++ wasm/Cargo.toml
@@ -11,3 +11,12 @@ categories.workspace = true
 
 [lib]
 crate-type = ["cdylib"]
+
+[features]
+debug = []
+
+[dependencies]
+console_error_panic_hook = "0.1.7"
+japanese-address-parser = { path = "../core" }
+wasm-bindgen = { workspace = true }
+wasm-bindgen-futures = { workspace = true }
